move pages around and delete dashboard page, just use index; redirect wasn't working
