--- packages/frontend/features/layout/PAGES.ts
@@ -1,8 +1,0 @@
-export const pages = [
-  { label: "Dashboard", id: "dashboard" },
-  { label: "Automate", id: "automate" },
-  { label: "Swap Tokens", id: "swap-tokens" },
-  { label: "Import Positions", id: "import-positions" },
-  { label: "Exit Positions", id: "exit-positions" },
-  { label: "Info", id: "info" },
-];
