+++ packages/frontend/pages/PAGES.ts
@@ -1,0 +1,12 @@
+export const pages = [
+  { label: "Dashboard", id: "dashboard", path: "/" },
+  { label: "Automate", id: "automate", path: "/automate" },
+  { label: "Swap Tokens", id: "swap-tokens", path: "/swap-tokens" },
+  {
+    label: "Import Positions",
+    id: "import-positions",
+    path: "/import-positions",
+  },
+  { label: "Exit Positions", id: "exit-positions", path: "/exit-positions" },
+  { label: "Info", id: "info", path: "/info" },
+];
