--- packages/frontend/pages/dashboard.tsx
@@ -1,13 +1,0 @@
-const Dashboard = () => {
-  return <>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-    <h1>I am the Dashboard!!!</h1>
-  </>
-};
-
-export default Dashboard;
